requirements.txt
streamlit==1.44.1
pandas==2.2.2
gspread==6.2.0
oauth2client==4.1.3
Pillow==10.2.0
numpy==1.26.4
